🎨style: 폴더 정리
