--- src/css/Main.css
@@ -1,192 +1,0 @@
-.mainHeader {
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  gap: 30px;
-  align-items: center;
-}
-
-.logo {
-  margin-top: 75px;
-}
-
-.search {
-  position: relative;
-}
-
-.search input {
-  width: 230px;
-  height: 36px;
-  border-radius: 10px;
-  border: 2px solid #3C6255;
-  padding-left: 10px;
-  padding-right: 35px;
-  margin-bottom: 40px;
-  outline: none;
-  font-size: 15px;
-}
-
-.searchButton {
-  all: unset;
-  position: absolute;
-  top: 10px;
-  right: 10px;
-  cursor: pointer;
-}
-
-.searchButton img {
-  width: 20px;
-  height: 20px;
-}
-
-.main {
-  margin-bottom: 80px;
-}
-
-.select {
-  display: flex;
-  justify-content: center;
-}
-
-.foodMenu {
-  display: flex;
-  width: 330px;
-  height: 50px;
-  margin-bottom: 40px;
-  justify-content: space-between;
-  align-items: center;
-  font-size: 15px;
-  font-family: 'NanumBarunGothicLight';
-}
-
-.foodMenu button {
-  all: unset;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-}
-
-.foodMenu button img {
-  margin-bottom: 10px;
-}
-
-.sushi button img {
-  margin-top: 3px;
-  margin-bottom: 13px;
-}
-
-.bannerImg {
-  display: flex;
-  align-items: center;
-  justify-self: center;
-  width: 390px;
-  height: 109px;
-  overflow: hidden;
-}
-
-.carousel {
-  display: flex;
-  transition: all 0.5s;
-}
-
-.carousel button {
-  all: unset;
-}
-
-.lrButton {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  gap: 20px;
-  margin-bottom: 40px;
-}
-
-.lrButton button {
-  all: unset;
-  padding: 5px;
-  cursor: pointer;
-}
-
-.storeButton {
-  all: unset;
-}
-
-.saleStores {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-}
-
-.saleWrite {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  margin-right: 220px;
-  margin-bottom: 20px;
-  gap: 10px;
-  font-size: 15px;
-  font-family: 'NanumBarunGothicBold';
-}
-
-.storeImg {
-  display: flex;
-  align-items: center;
-  justify-self: center;
-  margin-bottom: 10px;
-  border-radius: 10px;
-}
-
-.storeName {
-  margin-left: 10px;
-  margin-top: 15px;
-  margin-bottom: 2px;
-  font-size: 15px;
-  font-family: 'NanumBarunGothicBold';
-}
-
-.storeInfo {
-  display: flex;
-  gap: 10px;
-  margin-left: 10px;
-  font-size: 13px;
-}
-
-.oneStore {
-  margin-bottom: 20px;
-}
-
-/* body, footer 경계선 */
-.footerButton img {
-  margin-top: 10px;
-}
-
-/* footer 부분 버튼 배치 */
-.footerButton {
-  position: fixed;
-  bottom: -50px;
-  /* 화면의 아래쪽에 고정 */
-  left: 50%;
-  /* 가로 중앙에 위치 */
-  transform: translateX(-50%);
-  /* 버튼을 정확히 중앙으로 이동 */
-  cursor: pointer;
-  display: flex;
-  flex-direction: row;
-  justify-content: center;
-  border-top: 1px solid grey;
-}
-
-/* footer 버튼 배치 */
-.footerButton button {
-  width: 78px;
-  height: 82px;
-  border: white solid;
-  background-color: white;
-}
-
-/* footer 부분 버튼 */
-.footerbtnimg {
-  margin-bottom: 80px;
-}
